convert to creditmanager
